Extend cleanup timeout until US Labor Day
This intentionally sets a time bomb that will put the timeout back to 45
minutes in September.  There are other timeouts (90 minutes for a total
test), but this is the first we're hitting.  The current 45 minutes
cleanup time is a serious bug and we're struggling with debugging our
first real bug spanning components.  We need to free up progress on
other tests.